# back/utils/Swagger.yml
openapi: 3.0.3
info:
  title: Leitner system
  description: "This API aims to provide features to manage a graphical interface for Leitner System."
  version: 1.0.0
servers:
  - url: http://localhost:3000/docs
    description: Local server
paths:
  /login:
    post:
      tags:
        - Authentication
      summary: User Login
      description: Used to authenticate a user and obtain a token.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                  token:
                    type: string
        '401':
          description: Invalid credentials

  /register:
    post:
      tags:
        - Authentication
      summary: User Registration
      description: Used to register a new user.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        '400':
          description: Bad request

  /cards:
    get:
      tags:
        - Cards
      summary: Get all cards
      description: Used to fetch every cards with given tags. If no tags are provided, will fetch all cards.
      parameters:
        - in: query
          name: tags
          schema:
            type: array
            items:
              type: string
          description: Tags of cards to find. If not present, all cards will be found.
          example: tag1,tag2
      responses:
        '200':
          description: Found cards by tag query
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Card"
      security:
        - jwt: []
    post:
      tags:
        - Cards
      summary: Create a card
      description: Used to create a new card in the system. A new card will be present in the next quiz.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardUserData"
      responses:
        '201':
          description: Created card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        '400':
          description: Bad request
      security:
        - jwt: []
  /cards/quizz:
    get:
      tags:
        - Learning
      summary: Cards for the day
      description: Used to fetch all cards for a quiz at a given date. If no date is provided, the quiz will be for today.
      parameters:
        - in: query
          name: date
          description: Date of quiz. If not provided, the date will be today.
          example: 2023-11-03
          schema:
            type: string
            format: date
      responses:
        '200':
          description: All cards of quiz for today
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Card"
      security:
        - jwt: []

  /cards/{cardId}/answer:
    patch:
      tags:
        - Learning
      summary: Answer a question
      description: Used to answer a question. The body indicates if the user has answered correctly or not.
      parameters:
        - in: path
          name: cardId
          required: true
          description: Id of answered card.
          schema:
            $ref: "#/components/schemas/CardId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: ["isValid"]
              properties:
                isValid:
                  type: boolean
                  description: True if the user has answered correctly, false otherwise
      responses:
        '204':
          description: Answer has been taken into account
        '400':
          description: Bad request
        '404':
          description: Card not found
      security:
        - jwt: []

components:
  schemas:
    CardId:
      type: string
      description: Generated identifier of a card
      example: 6c10ad48-2bb8-4e2e-900a-21d62c00c07b
    CardUserData:
      type: object
      required: ["question", "answer"]
      properties:
        question:
          type: string
          description: Question to be asked to the user during a quiz
          example: "What is pair programming?"
        answer:
          type: string
          description: Expected answer for the question
          example: "A practice to work in pair on the same computer."
        tag:
          type: string
          description: A tag to group cards on the same topic
          example: "Teamwork"
    Card:
      allOf:
        - type: object
          required: ["id", "category"]
          properties:
            id:
              $ref: "#/components/schemas/CardId"
            category:
              $ref: "#/components/schemas/Category"
        - $ref: "#/components/schemas/CardUserData"
    Category:
      type: string
      description: Category of card indicating how many times you answered it and appearance frequency
      example: FIRST
      enum:
        - FIRST
        - SECOND
        - THIRD
        - FOURTH
        - FIFTH
        - SIXTH
        - SEVENTH
        - DONE

    LoginRequest:
      type: object
      required: ["email", "password"]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    RegisterRequest:
      type: object
      required: ["email", "password", "firstname", "lastname"]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        firstname:
          type: string
        lastname:
          type: string

    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstname:
          type: string
        lastname:
          type: string